refactor(server): move project og resolution to layout components
