+++ server/lib/tuist_web/components/layout_components.ex
@@ -6,7 +6,7 @@ defmodule TuistWeb.LayoutComponents do
 
   import TuistWeb.CSP, only: [get_csp_nonce: 0]
 
-  alias TuistWeb.Helpers.OpenGraph
+  alias Tuist.Projects.OpenGraph, as: ProjectsOpenGraph
 
   attr(:current_user, :map, default: nil)
 
@@ -53,7 +53,7 @@ defmodule TuistWeb.LayoutComponents do
   end
 
   def head_meta_meta_tags(assigns) do
-    resolved_head_image = OpenGraph.resolved_head_image(assigns)
+    resolved_head_image = resolved_head_image(assigns)
     assigns = assign(assigns, :resolved_head_image, resolved_head_image)
 
     ~H"""
@@ -78,8 +78,8 @@ defmodule TuistWeb.LayoutComponents do
   end
 
   def head_x_meta_tags(assigns) do
-    resolved_head_image = OpenGraph.resolved_head_image(assigns)
-    resolved_twitter_card = OpenGraph.resolved_twitter_card(assigns)
+    resolved_head_image = resolved_head_image(assigns)
+    resolved_twitter_card = resolved_twitter_card(assigns)
 
     assigns =
       assigns
@@ -196,4 +196,67 @@ defmodule TuistWeb.LayoutComponents do
   end
 
   defp maybe_add_group(groups, _group_type, _entity), do: groups
+
+  defp resolved_head_image(assigns) do
+    if is_binary(assigns[:head_image]) do
+      assigns[:head_image]
+    else
+      public_project_head_image(assigns)
+    end
+  end
+
+  defp resolved_twitter_card(assigns) do
+    if is_binary(assigns[:head_twitter_card]) do
+      assigns[:head_twitter_card]
+    else
+      default_twitter_card(assigns)
+    end
+  end
+
+  defp public_project_head_image(assigns) do
+    with %{visibility: :public, name: project_handle} = project <- assigns[:selected_project],
+         %{name: account_handle} = account <- assigns[:selected_account] do
+      title = ProjectsOpenGraph.default_title(assigns[:head_title])
+      key_values = resolve_key_values(assigns, account, project)
+
+      ProjectsOpenGraph.image_url(account_handle, project_handle, title, key_values)
+    else
+      _ ->
+        Tuist.Environment.app_url(path: "/images/open-graph/card.jpeg")
+    end
+  end
+
+  defp default_twitter_card(assigns) do
+    if public_project?(assigns) do
+      "summary_large_image"
+    else
+      "summary"
+    end
+  end
+
+  defp public_project?(assigns) do
+    case assigns[:selected_project] do
+      %{visibility: :public} -> true
+      _ -> false
+    end
+  end
+
+  defp resolve_key_values(assigns, account, project) do
+    custom_key_values =
+      assigns
+      |> Map.get(:head_open_graph_key_values, [])
+      |> List.wrap()
+
+    default_key_values = ProjectsOpenGraph.default_key_values(account, project)
+
+    key_values =
+      if custom_key_values == [] do
+        page_title = ProjectsOpenGraph.default_title(assigns[:head_title])
+        [%{key: "Page", value: page_title} | default_key_values]
+      else
+        custom_key_values ++ default_key_values
+      end
+
+    Enum.take(key_values, 3)
+  end
 end
